# services/stash-service/api-spec/src/main/resources/openapi/stash-api.yaml
openapi: 3.0.3
info:
  title: Linkpouch Stash Service API
  description: API for managing stashes and links in Linkpouch
  version: 1.0.0
  contact:
    name: Linkpouch Team
servers:
  - url: /api
    description: Local development
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT issued by POST /stashes/{stashId}/access-token
  schemas:
    CreateStashRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Name of the stash
        password:
          type: string
          minLength: 1
          maxLength: 255
          description: Optional passphrase to protect the stash
    AcquireAccessRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 255
          description: Required when the stash is password-protected
    AccessTokenResponse:
      type: object
      required:
        - accessToken
        - expiresIn
        - isClaimer
        - isStashClaimed
      properties:
        accessToken:
          type: string
          description: Short-lived JWT to use as Bearer token on all stash/link endpoints
        expiresIn:
          type: integer
          description: Token lifetime in seconds
        isClaimer:
          type: boolean
          description: True if the token was issued to the account that claimed this stash
        isStashClaimed:
          type: boolean
          description: True if any account has claimed this stash
    SetPasswordRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 255
          description: New passphrase to protect the stash
    StashResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        passwordProtected:
          type: boolean
          description: Whether this stash requires a password to access
        visibility:
          type: string
          enum:
            - SHARED
            - PRIVATE
          description: Who can access this stash (SHARED = anyone with URL, PRIVATE = claimer only)
        linkPermissions:
          type: string
          enum:
            - FULL
            - READ_ONLY
          description: Write access for non-owner visitors (FULL = anyone can edit, READ_ONLY = owner only)
        linkCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        signedUrl:
          type: string
          description: Signed URL for accessing this stash (only returned on creation)
...
    UpdateStashRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: New name for the stash
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        errorCode:
          type: string
          description: Machine-readable error code (e.g. PASSWORD_REQUIRED, INVALID_TOKEN)
        path:
          type: string
        signatureRefreshedAt:
          type: string
          format: date-time
          description: When the stash signature was last regenerated (only set when errorCode is SIGNATURE_REGENERATED)
    SseTicketResponse:
      type: object
      required:
        - ticket
        - expiresIn
      properties:
        ticket:
          type: string
        expiresIn:
          type: integer
          description: Ticket lifetime in seconds
...
paths:
  /stashes:
    post:
      summary: Create a new stash
      operationId: createStash
      tags:
        - Stashes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStashRequest'
      responses:
        '201':
          description: Stash created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/access-token:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Stash-Signature
        in: header
        required: false
        schema:
          type: string
        description: >
          HMAC-SHA256 signature proving URL access. Required for non-private stashes.
          Omit when using an account JWT to access a private stash as the claimer.
    post:
      summary: Acquire a short-lived access token for this stash
      description: >
        Unified token acquisition endpoint. Accepts three optional credentials:
        X-Stash-Signature (required for non-private stashes), Authorization: Bearer <accountJwt>
        (optional; used to detect claimer and grant access to private stashes), and password in the
        request body (required when the stash is password-protected).
        Returns 401 with errorCode PASSWORD_REQUIRED when the password is missing.
        Returns 401 with errorCode TOKEN_VERSION_MISMATCH when an existing token is stale.
      operationId: acquireAccessToken
      tags:
        - Stashes
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcquireAccessRequest'
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '401':
          description: Invalid signature, wrong password, or missing password (errorCode PASSWORD_REQUIRED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Stash is private and caller is not the claimer (errorCode STASH_PRIVATE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get stash by ID
      operationId: getStash
      tags:
        - Stashes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stash found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update stash name
      operationId: updateStash
      tags:
        - Stashes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStashRequest'
      responses:
        '200':
          description: Stash updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete stash by ID
      operationId: deleteStash
      tags:
        - Stashes
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Stash deleted successfully
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/password:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Set or change the stash password
      description: >
        Sets or changes the password on this stash.
        Requires a claimer Bearer token (proof of ownership).
        If the stash is currently password-protected, also requires a valid Bearer token
        with a matching pwdKey to prove knowledge of the current password.
      operationId: setStashPassword
      tags:
        - Stashes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      responses:
        '200':
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid signature or invalid Bearer token for currently protected stash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove the stash password
      description: >
        Removes password protection from this stash.
        Requires the stash signature (proof of ownership) and a valid Bearer token
        with a matching pwdKey to prove knowledge of the current password.
      operationId: removeStashPassword
      tags:
        - Stashes
      responses:
        '204':
          description: Password removed successfully
        '401':
          description: Invalid signature or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/regenerate-signature:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Regenerate the stash signature (invalidates all old shared URLs)
      description: >
        Generates a new secret key for the stash, invalidating the previous signed URL.
        Only the claimer (or unclaimed stash owner) may call this.
        Returns the stash with the new signedUrl in the response.
      operationId: regenerateSignature
      tags:
        - Stashes
      responses:
        '200':
          description: Signature regenerated — new signedUrl in the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StashResponse'
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the owner can perform this action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stash not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/sse-ticket:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Issue a short-lived SSE ticket for the stash
      description: >
        Clients exchange their Bearer stash-access token for a short-lived ticket before
        opening an SSE connection. The ticket is passed as a query parameter so that the
        Bearer token never appears in server access logs.
      operationId: issueSseTicket
      tags:
        - Stashes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ticket issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SseTicketResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...